chore: Run Dependabot monthly
As this repository does not run in a context that is accessible to users or external requests, the risk of running old versions of dependencies for a short time is minimal. Update Dependabot's configuration to run monthly to reduce the frequency of PRs by a quarter and thus require less commitment from the team.